Fixed bug with LAST, added test script
- Fixed typo for image validation failure output
- Improved debug printing for conditional loops
- Writes to LAST now correctly write to the last port
- Added test_saves.sh fish script

test_saves.sh will run the simulator on every save in the provided
directory that has a filename matching the level ID. It has several
assumptions about the way I organize my saves built into it, so it may
not be useful for anyone else without modification. It currently runs
every level that I've implemented random tests for. (the first campaign,
including secret level, and the two image test patterns in the second
campaign, which have no input and therefore are randomness-invariant).