Refactor logging, yet again
Runtime in notice improves by a factor of 2.
The time was wasted creating null loggers in the ostream initialization.
So use composition instead of inheritance and never initialize the ostream insterface if you're a null logger

Logger loses the ostream interface, that's easily fixed with a couple templates.
No changes to debug logging.

The time spent logging in notice is finally negligible.