Added on-demand progress reporting
Added:
SIGUSR1 and SIGUSR2 no longer abort the sim, instead they, along with
SIGTSTP (sent by ^Z from terminal) now log_info() a message similar to
the one emitted when the seeds are exhausted and simming is complete.
SIGTSTP will still cause the sim to stop after the message is logged,
which does mean that at least one thread much complete simming a seed
and return to the main loop, so a slight delay is possible. (Feature not
implemented on Windows.)

I did check, because it's a bit ugly, but there seemingly is no better
way to respond to SIGTSTP than to raise(SIGSTOP) once ready to suspend.